[templating] rewrite pagination helpers
join pagination helpers to helper collection
