+++ library/umi/templating/extension/helper/type/pagination/PaginationHelper.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * UMI.Framework (http://umi-framework.ru/)
+ *
+ * @link      http://github.com/Umisoft/framework for the canonical source repository
+ * @copyright Copyright (c) 2007-2013 Umisoft ltd. (http://umisoft.ru/)
+ * @license   http://umi-framework.ru/license/bsd-3 BSD-3 License
+ */
+namespace umi\templating\extension\helper\type\pagination;
+
+/**
+ * Помощник вида для отображения постраничной навигации.
+ */
+class PaginationHelper
+{
+    /**
+     * Возвращает коллекцию помощников для вывода пагинации.
+     * @return PaginationHelperCollection
+     */
+    public function __invoke()
+    {
+        return new PaginationHelperCollection();
+    }
+}
